finished the products test alpha

+++ priv/repo/migrations/20230519201939_create_products.exs
@@ -6,7 +6,7 @@ defmodule Gojo.Repo.Migrations.CreateProducts do
       add :title, :string
       add :description, :string
       add :price, :decimal, precision: 15, scale: 6, null: false
-      add :sku, :integer, null: false
+      add :sku, :bigint, null: false
       add :user_id, references(:users, on_delete: :delete_all), null: false
 
       # quantity, image url, category, manufacturer.
